src/main/java/com/pragmaticcoders/checkout/view/ItemView.java
package com.pragmaticcoders.checkout.view;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Getter
@AllArgsConstructor
@NoArgsConstructor(access = AccessLevel.PROTECTED)
public class ItemView {
    private UUID id;
    private String name;
    private List<Price> prices;

    @Getter
    @AllArgsConstructor
    @NoArgsConstructor(access = AccessLevel.PROTECTED)
    public static class Price {
        private Integer quantity;
        private Integer price;
    }
}
